The basics
The basic structure of the main class and preprocess class.

Ideally, a third class will come along which will handle the
detection of the box. For all I care this can be called the vision
class.

The main class will load in the desired image, and then call functions
from the classes to perform operations on the image.